boiler_buy_api/requirements.txt
django
djangorestframework
django-cors-headers
drf-yasg
requests
psycopg2-binary
python-dotenv
whitenoise
pillow

